Move shared ui components into 'components/shared' folder
