+++ src/components/shared/Banner/Banner.js
