+++ src/components/shared/Banner/Banner.test.js
