+++ src/components/shared/FontAwesome/FontAwesome.js
