+++ src/components/shared/FontAwesome/FontAwesome.test.js
