+++ src/components/shared/InputGroup/InputGroup.test.js
