+++ src/components/shared/MainImage/MainImage.css
