+++ src/components/shared/MainImage/MainImage.js
