+++ src/components/shared/MainImage/MainImage.test.js
