+++ src/components/shared/MiniButton/MiniButton.js
