+++ src/components/shared/MiniButton/MiniButton.test.js
