+++ src/components/shared/Pill/Pill.js
