+++ src/components/shared/ReceiptItem/ReceiptItem.css
