+++ src/components/shared/Slider/Slider.css
