+++ src/components/shared/Slider/Slider.js
