+++ src/components/shared/Slider/Slider.test.js
